requirements.txt
ttkbootstrap==1.10.1
Pillow==10.0.0
pandas==2.0.3
openpyxl==3.1.2
requests==2.31.0
babel==2.13.1
pyinstaller==6.1.0